pyproject.toml
[tool.poetry]
name = "repl-nix-stellar"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "repl_nix_stellar"}]

[tool.poetry.dependencies]
python = "^3.12"
aiohttp = "^3.9.5"
discord-py = "^2.4.0"
requests = "^2.32.3"
flask = "^3.0.3"
asyncpg = "^0.29.0"
loguru = "^0.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
